Made levels into a service
